Chonk diff. Added observer pattern and ability to click on squares.

+++ src/gfx/EBO.cpp
@@ -1,6 +1,6 @@
 #include "EBO.hpp"
 
-EBO::EBO(const int* indices, unsigned int size) {
+EBO::EBO(const int* indices, size_t size) {
     glc(glGenBuffers(1, &id));
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, id);
     glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(int) * size, indices, GL_STATIC_DRAW);
